requirements.txt
fastapi
uvicorn
jinja2
numpy
pillow
tensorflow
gdown
python-multipart
scikit-learn
